Use composition instead of inheritance
