+++ src/main/java/org/testng/internal/ExpectedExceptionsHolder.java
@@ -1,60 +1,107 @@
 package org.testng.internal;
 
+import org.testng.IExpectedExceptionsHolder;
 import org.testng.ITestNGMethod;
+import org.testng.TestException;
 import org.testng.annotations.IExpectedExceptionsAnnotation;
 import org.testng.annotations.ITestAnnotation;
 import org.testng.internal.annotations.IAnnotationFinder;
 
-import java.util.regex.Pattern;
+import java.util.Arrays;
 
-/**
- * A class that contains the expected exceptions and the message regular expression.
- * @author cbeust
- */
-public class ExpectedExceptionsHolder extends AbstractExpectedExceptionsHolder {
-  public static final String DEFAULT_REGEXP = ".*";
+public class ExpectedExceptionsHolder {
 
-  public ExpectedExceptionsHolder(IAnnotationFinder finder, ITestNGMethod method) {
-    super(finder, method);
+  protected final IAnnotationFinder finder;
+  protected final ITestNGMethod method;
+  private final Class<?>[] expectedClasses;
+  private final IExpectedExceptionsHolder holder;
+
+  protected ExpectedExceptionsHolder(IAnnotationFinder finder, ITestNGMethod method, IExpectedExceptionsHolder holder) {
+    this.finder = finder;
+    this.method = method;
+    expectedClasses = findExpectedClasses(finder, method);
+    this.holder = holder;
   }
 
-  /**
-   *   message / regEx  .*      other
-   *   null             true    false
-   *   non-null         true    match
-   */
-  @Override
-  protected boolean isExceptionMatches(Throwable ite) {
-    String messageRegExp = getRegExp();
-
-    if (DEFAULT_REGEXP.equals(messageRegExp)) {
-      return true;
-    }
-
-    final String message = ite.getMessage();
-    return message != null && Pattern.compile(messageRegExp, Pattern.DOTALL).matcher(message).matches();
-  }
-
-  protected String getWrongExceptionMessage(Throwable ite) {
-    return "The exception was thrown with the wrong message:" +
-           " expected \"" + getRegExp() + "\"" +
-           " but got \"" + ite.getMessage() + "\"";
-  }
-
-  private String getRegExp() {
+  private static Class<?>[] findExpectedClasses(IAnnotationFinder finder, ITestNGMethod method) {
     IExpectedExceptionsAnnotation expectedExceptions =
         finder.findAnnotation(method, IExpectedExceptionsAnnotation.class);
+    // Old syntax
     if (expectedExceptions != null) {
-      // Old syntax => default value
-      return DEFAULT_REGEXP;
+      return expectedExceptions.getValue();
     }
 
     // New syntax
     ITestAnnotation testAnnotation = finder.findAnnotation(method, ITestAnnotation.class);
     if (testAnnotation != null) {
-      return testAnnotation.getExpectedExceptionsMessageRegExp();
+      return testAnnotation.getExpectedExceptions();
     }
 
-    return DEFAULT_REGEXP;
+    return new Class<?>[0];
+  }
+
+  /**
+   * @param ite The exception that was just thrown
+   * @return true if the exception that was just thrown is part of the
+   * expected exceptions
+   */
+  public boolean isExpectedException(Throwable ite) {
+    if (!hasExpectedClasses()) {
+      return false;
+    }
+
+    // TestException is the wrapper exception that TestNG will be throwing when an exception was
+    // expected but not thrown
+    if (ite.getClass() == TestException.class) {
+      return false;
+    }
+
+    Class<?> realExceptionClass= ite.getClass();
+
+    for (Class<?> exception : expectedClasses) {
+      if (exception.isAssignableFrom(realExceptionClass) && holder.isThrowableMatching(ite)) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+
+  public Throwable wrongException(Throwable ite) {
+    if (!hasExpectedClasses()) {
+      return ite;
+    }
+
+    if (holder.isThrowableMatching(ite)) {
+      return new TestException("Expected exception of " +
+                               getExpectedExceptionsPluralize()
+                               + " but got " + ite, ite);
+    } else {
+      return new TestException(holder.getWrongExceptionMessage(ite), ite);
+    }
+  }
+
+  public TestException noException(ITestNGMethod testMethod) {
+    if (!hasExpectedClasses()) {
+      return null;
+    }
+    return new TestException("Method " + testMethod + " should have thrown an exception of "
+                             + getExpectedExceptionsPluralize());
+  }
+
+  private boolean hasExpectedClasses() {
+    return expectedClasses != null && expectedClasses.length > 0;
+  }
+
+  private String getExpectedExceptionsPluralize() {
+    StringBuilder sb = new StringBuilder();
+    if (expectedClasses.length > 1) {
+      sb.append("any of types ");
+      sb.append(Arrays.toString(expectedClasses));
+    } else {
+      sb.append("type ");
+      sb.append(expectedClasses[0]);
+    }
+    return sb.toString();
   }
 }
